# openapi.yml
openapi: 3.1.0
info:
  title: Masashi Enokida API
  version: 1.0.0
servers:
  - url: http://localhost:4000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
...
    Concert:
      type: object
      required: [id, title, date, venue, is_upcoming]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        date:
          type: string
          format: date-time
        time:
          type: string
          nullable: true
        venue:
          type: string
        address:
          type: string
          nullable: true
        image_url:
          type: string
          nullable: true
        program:
          type: array
          items:
            type: string
          nullable: true
        price:
          type: string
          nullable: true
        ticket_url:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
        is_upcoming:
          type: boolean
        created_at:
          type: string
          format: date-time
    ConcertInput:
      type: object
      required: [title, date, venue]
      properties:
        title:
          type: string
        date:
          type: string
          format: date-time
        time:
          type: string
          nullable: true
        venue:
          type: string
        address:
          type: string
          nullable: true
        image_url:
          type: string
          nullable: true
        program:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          nullable: true
        price:
          type: string
          nullable: true
        ticket_url:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
        is_upcoming:
          type: boolean
          default: true
        is_published:
          type: boolean
          default: true
...
    BiographyEntry:
      type: object
      required: [id, year, description, sort_order]
      properties:
        id:
          type: string
          format: uuid
        year:
          type: string
        description:
          type: string
        sort_order:
          type: integer
...
paths:
...
  /concerts:
    get:
      operationId: getConcerts
      summary: List concerts
      parameters:
        - name: upcoming
          in: query
          schema:
            type: string
            enum: ['true', 'false']
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Concert'
    post:
      operationId: createConcert
      summary: Create concert (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConcertInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Concert'
  /concerts/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getConcertById
      summary: Get single concert
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Concert'
        '404':
          description: Not found
    put:
      operationId: updateConcert
      summary: Update concert (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConcertInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Concert'
    delete:
      operationId: deleteConcert
      summary: Delete concert (admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
  /discography:
    get:
      operationId: getDiscography
      summary: List discography
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DiscographyItem'
    post:
      operationId: createDiscography
      summary: Create discography item (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscographyInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscographyItem'
  /discography/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getDiscographyById
      summary: Get single discography item
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscographyItem'
        '404':
          description: Not found
    put:
      operationId: updateDiscography
      summary: Update discography item (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscographyInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscographyItem'
    delete:
      operationId: deleteDiscography
      summary: Delete discography item (admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
  /biography:
    get:
      operationId: getBiography
      summary: List biography entries
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BiographyEntry'
    post:
      operationId: createBiography
      summary: Create biography entry (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BiographyInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BiographyEntry'
  /biography/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    put:
      operationId: updateBiography
      summary: Update biography entry (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BiographyInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BiographyEntry'
    delete:
      operationId: deleteBiography
      summary: Delete biography entry (admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
...
  /blog:
    get:
      operationId: getBlog
      summary: List blog posts (with member gating)
      security:
        - {}
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: category
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlogPostList'
  /blog/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getBlogPost
      summary: Get single blog post (with member gating)
      security:
        - {}
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlogPost'
        '404':
          description: Not found
...
  /admin/stats:
    get:
      operationId: getAdminStats
      summary: Get admin dashboard stats
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminStats'
  /admin/contacts:
    get:
      operationId: getAdminContacts
      summary: List contacts (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [unread, read, replied, archived, all]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminContactList'
  /admin/contacts/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    put:
      operationId: updateAdminContact
      summary: Update contact status (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactStatusUpdate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminContact'
        '404':
          description: Not found
  /admin/members:
    get:
      operationId: getAdminMembers
      summary: List members (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [USER, MEMBER_FREE, MEMBER_GOLD, ADMIN, all]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminMemberList'
  /admin/concerts:
    get:
      operationId: getAdminConcerts
      summary: List all concerts including unpublished (admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminConcert'
  /admin/concerts/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getAdminConcertById
      summary: Get single concert (admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminConcert'
        '404':
          description: Not found
  /admin/discography:
    get:
      operationId: getAdminDiscography
      summary: List all discography including unpublished (admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminDiscographyItem'
  /admin/discography/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getAdminDiscographyById
      summary: Get single discography item (admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDiscographyItem'
        '404':
          description: Not found
  /admin/biography:
    get:
      operationId: getAdminBiography
      summary: List all biography entries (admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminBiographyEntry'
  /admin/biography/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getAdminBiographyById
      summary: Get single biography entry (admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminBiographyEntry'
        '404':
          description: Not found
  /admin/blog/categories:
    get:
      operationId: getAdminBlogCategories
      summary: List blog categories (admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminBlogCategory'
    post:
      operationId: createAdminBlogCategory
      summary: Create blog category (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminBlogCategoryInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminBlogCategory'
        '409':
          description: Conflict (duplicate name or slug)
  /admin/blog/categories/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    put:
      operationId: updateAdminBlogCategory
      summary: Update blog category (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminBlogCategoryInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminBlogCategory'
        '404':
          description: Not found
        '409':
          description: Conflict
    delete:
      operationId: deleteAdminBlogCategory
      summary: Delete blog category (admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '404':
          description: Not found
        '409':
          description: Conflict (category has associated posts)
...